fixed bug in the positioning of the voxels for the EndCap Forward
The size of the voxel_rotation array in 'dream.py' increased from 2x3 to 3x3 in order to allow for an extra rotation of 90 degrees around the Z-axis for the voxels of the Forward EndCap. The values of the 3 new rotation angles added to the voxel_rotation array are read from the output generated by running the scripts 'DREAMSUMOx_calculate_voxels.py' (the new values are inserted at the end of the 'tempF3' string, part of the 'fF' output file).